requirements.txt
asgiref==3.5.2
astroid==2.11.7
boto3==1.24.36
botocore==1.27.36
certifi==2022.6.15
dill==0.3.5.1
Django==4.0.6
django-elasticsearch-dsl==7.2.2
django-environ==0.9.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-nested-routers==0.93.4
elastic-transport==8.1.2
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0
isort==5.10.1
jmespath==1.0.1
lazy-object-proxy==1.7.1
Markdown==3.3.7
mccabe==0.7.0
mysqlclient==2.1.1
platformdirs==2.5.2
PyJWT==2.4.0
pylint==2.14.4
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.1
urllib3==1.26.10
uWSGI==2.0.20
wrapt==1.14.1
